Add body skeleton
This adds basic responsive layout to the main content section of the
page - things like the article headline, meta data, main media, and
article body.

It's a lightweight pass; there is obviously a lot of work still to be
done. But attempting to bundle everything into a single PR is probably
not a good strategy. Once this is merged, people can potentially work
in parallel on what remains. But either way, getting early
feedback/smaller PRs seems better.