Initial GAP commit
GAP stands for Guardian Accelerated Pages. GAP is inspired by AMP
but aims to be less strict (so that it can be used on existing
production sites with less effort), and also easier to extend
(extensions don't need to be officially adopted). It should still
offer some good out-of-the-box performance though when used
sensibly.