Refactor commercialProperties as map of targeting params
A refactor of commercial properities passed to AMP ads. Previously
this was an array of params, and we had to filter the array for the
correct edition.

This change restructures commercial properties as a map of edition
to targeting params, making it much easier to select the correct
params for the current edition. It simplifies our code and makes
it more efficient.

This corresponds to guardian/frontend#21208. We can remove the original
commercialProperties after this change is merged.

This provides the groundwork for paid content theming, as Branding is
easier to retrieve from the new editionalised map.