Tidy AMP packages
* remove components/lib
* remove components/primitives

The latter does reduce some type safety. However, we don't use the
sole component (Img) for most of our amp images and it's a lot of
code for only a small amount of benefit (it's not too hard given
the fast feedback to work out if you've mispelt an attribute).