Inline Guui
By this is meant, breaking the Guui dependency of the frontend
package by copying all of the used content into it.

This way we can then remove Guui.

I've kept Guui around for now as it's used by the design package
at the moment so we can't outright delete it.