Handle unknown element types without blowing up
We want to accept unknown element types as it makes adding new ones
easier and also is easier on clients.

Unfortunately, this complicates things because we also want to be
typesafe. Previously we weren't.

I'm not overly-zealous on type safety here. This solution works
but is a bit verbose/unclear.

Alternatives are:

* use the new GenericElement on CAPIType but immediately coerce
into a stricter type
* only send supported types from Frontend

None of these seem ideal!