Update typography to src-foundations
The mappings between pasteup and src-foundations are not quite 1-2-1.

To start with I've been naive and coerced things when no perfect
equivalent font-size existed.

I've also not customised line-heights yet, so things are likely to
be wrong.

But it's a starting point anyway.

The mappings are:

textSans (pasteup -> src-foundations):
  1 : 1
  2..4 : 2
  5..6 : 3
  7..8 : 4
  9 : 5

body:

  1 : 1
  2..4 : 2

headline:

1..2 : 1
3 : 2
4 : 3
5 : 4
6..7 : 5
8..9 : 6