Fix local fonts
I’d made this change to the dev server to be consistent with the change
I made to the production server, but it seems to have caused an issue.
(I tested locally, and thought it was all fine, but must have not re-
started the server so got a false positive).