Implement Picnic ad slot recommendations
Add ad slot format details as per Picnic's recommendations. The result
of these changes is that we will add the `data-multi-size` property to
ad tags in AMP and increase the primary ad size to be 320x480.

This will have the effect of making the primary ad slot larger than the
current size of 300x250. If the new larger ad slot cannot be filled
with an ad of that size there are two resulting situations:

1. If the ad renders *below the viewport*, the runtime automatically
resizes the ad below the line, with not visual reflow for the user.

2. If the ad renders *inside the viewport*, the runtime will centre the
ad in the middle of the reserved ad space.

Given that picnic is intended to serve ads mid-article, they should
appear below the fold meaning that situation 1) is the most likely.

Information about `multi-data-size` can be found in:
https://medium.com/ampfuel/increase-ad-competition-with-multi-size-fluid-ads-in-amp-9cff90066c16