Reformat for A/B testing with amp-experiment
This separates out the picnic ad style and component from the existing
ad component and style. This is a pre-step before implementing A/B
testing in AMP which will require both style to be available and
in separate components.

This also places the two ad styles in to separate `div` elements to
allow one of them to be turned off using the `amp-experiment` method of
CSS styling variants.