Add amp-experiment component
In order to perform proper A/B testing in AMP we need to use the
`amp-experiment` tag to create variant groups.
This commit adds this component and places it in the Article component,
test document and `render.tsx`.

It does not update the analytics to send the variant information as
this will require updating the ophan config information.