Integrate Braze Web SDK to show banner messages
Use the Braze Web SDK to determine if there is a message to show the
user. The SDK is lazily loaded. At the moment only logged in digipack
subscribers will load the SDK since they're the only users targeted by
our initial campaign. If there is a message to show, lazily load the
banner component and render it.

Note the the banner is not ready and the campaign isn't configured in
Braze yet so this feature will be hidden behind a switch.