Introduce banner picker to DCR
This behaves very similarly to the banner picker on frontend. Banners
are asked concurrently if they can show, and the highest priority banner
which answers "yes" wins.

We support configurable timeouts around the canShow logic. The timeout
has been applied to the reader revenue banner, but not the CMP. Timeouts
are recorded in Ophan.