Fix issue where we weren’t running banner picker
The brazeUuid is a dependency for the banner picking logic, so we want
to make sure it’s always defined. We also want to make sure once the
banner picker’s dependencies are all available that we only run it once
(we don’t want to switch the banner once we’ve picked one to show).