Fix flip-flopping Cypress test
The cypress/integration/e2e/article.e2e.spec.js has been sporadically
failing recently. Looking at the Team City history it looks likely that
this started happening when the audience % for the curated container
test was increased in #1963. It's not totally clear why that's caused
the failure to sometimes happen (presumably when we fall into one of the
variants, not the control). Forcing the control via the URL seems to get
things behaving more consistently, so going with that as a fix for now.

I also slightly refactored the web e2e article tests so that each
article gets its own it block, since it's easier to see where failures
happened that way.