Ensure local SDK data is wiped on logout
If we’ve recorded in localstorage that there is a braze user, but we
haven’t got a brazeUuid, then we treat that as meaning that the user has
signed out, so tell the SDK to wipe data.

We have to initialize the SDK before wiping. Split out the apiKey and
braze switch condition, without which we can’t do anything, from the
preChecks. The preChecks are now more business level checks which will
probably change over time.