Fix card index marker when using buttons
Fix broken progress indicator when using buttons. Prior to this the
index was often off by 1, seemingly because the state was captured prior
to updating the carousel.

Here we have moved to using `useReducer` to dispatch events related to
particular actions. This means that we can capture desired and current
behaviour/state, making the progress indicator more reliable.

TODO:
- Capture scroll progress and update progress bar.
- Extract adjusted dots logic to be usable in both CSS and JS (depends
on viewport size).

Co-authored-by: nitro-marky <mark.addis@guardian.co.uk>