Making a start on ThirdPartyEmbeddedContent.tsx component
This is intended to be a wrapper around any embedded content displayed on the site.

It will make the decision if the content needs to be wrapped in a Click to View wrapper
based on the flag in the response from frontend. It also constructs the headline and
body text for ClickToView overlay, and attempts to supply the correct height and
width for the overlay.