Add support for the Braze Epic
This adds a second candidate, BrazeEpic, alongside ReaderRevenueEpic to
the end of article slot.

This is missing some tracking events, which we'll add before using the
Braze epic in production.

The wrapped Braze SDK instance is shared across SlotBodyEnd and
StickyBottomBanner keeping it as state in App and passing down as a
prop.

Co-authored-by: Tom Wey <tom.wey@theguardian.com>