Remove '!important' from atom bg CSS (#1226)
In a majority of cases atoms should have a white background in dark mode on apps. However we are now creating interactive atoms with a styleset that allows them to have dark mode styles which should be shown on the default dark background. 

In this atoms I want to be able to override the `background: white;` in the atom CSS. Only atoms where there are special dark mode styles will do this. Currently I can't override it because of the `!important` on the style rule here. 

The styles here run after the styles from the atom so if they are more powerful or have the same power these styles win out.