Add Cypress tests around Braze web functionality
This is an attempt to encode in a test a check we are otherwise doing
manually, which is that when a user logs out (having been eligible for
messages from Braze) we'll clear related data.

In order for Braze to run on the page we need the user's consent. In the
tests consent is given by interacting with the CMP banner, just like a
real user. Because the banner varies in different geo locations we
explicitly set the country code for consistent behaviour across
environments (e.g. local dev, CI).