Rename update manual fixtures code
It wasn't clear what the specific differences were between the data
generated/copied from manual JSON files and those that were generated
from `PROD` data.

I've updated this here so that we're clear that this is updating the
fixtures based on manual data. I've also moved the fixtures that are
generated from the data into the `generated/` directory to make it
clear, as well as moving the data itself into the `fixtures/` directory.