Fix analysis title styling on AMP
The headline in Analysis articles on AMP was missing the underline found
on the main site.

This adds a new `TopMeta...` component specific to Analysis. It's based
on `TopMetaNews` but with underlined styles added. This follows the
existing pattern of having separate `TopMeta...` components for each
specific supported `Design` type.