Decouple BrazeEpic from the RRCP component
This moves the BrazeEpic away from "borrowing" the ContributionsEpic
component to instead use components from @guardian/braze-components,
just like the banner. This is required because we want to start using
some new components to promote newsletters in this space. It was also
fragile to rely on a component from another team/system.

In @guardian/braze-components we've split the top level factory
component BrazeMessageComponent into two - BrazeBannerComponent and
BrazeEndOfArticle component.

Co-authored-by: Michael Clapham <michaeljclapham@gmail.com>