remove list and listtype from youtube block component
Youtube videos in amp were not playing because of invalid values set against
query params in the youtube iframe src. The list and listtype params were
incorrect, and are only needed to identify the video if we don't have the
videoId. Here, we believe it's safe to remove the data attributes as we
have the videoId.