Add `performance.now` to `polyfill.io` bundle
Some older browsers don't support the User Timing API [1]. Mostly we guard against this case by checking if `window.performance` is `undefined` before calling methods on it, but we discovered one instance where this wasn't happening and we believe this to be the cause of a number of Sentry errors. Adding a polyfill for older browsers feels like the right approach so we don't always need to worry about guards.

[1]: https://caniuse.com/user-timing