add to legacy interactive styling
The Visuals team highlighted some errors that were blocking the
migration of legacy interactives, which included an error making the
background colour extend the full width and the margins not aligned.
The background colour can be fixed by applying styling to the
<article> tag whereas to resolve the margin alignment I had to undo
some of the styling previously added. This means some legacy articles
will regress if rendering via DCR, see DCR PR #3284