Fix bug with documentaries placeholder
There were some pre-requisities to solving the placeholder bug for documentary
interactives. Firstly it was decided to add the InteractiveBlockElement as a
bareElement inside renderElement.tsx - it didn't quite make sense to wrap our
Figure inside a Figure. Secondly, it was decided that the InteractiveBlockComponent
should be a Portal instead of HydrateOnce, because on client-side we're inserting
content (replacing the server-rendered placeholder). Using Portal allowed
documentaries to overwrite content successfully.