Use HydrateOnce to remove placeholders.
InteractiveBlockComponents use hydration to ensure server-rendered content
is updated on the client. However, the server-rendered content uses placeholders
and the interactive can be loaded onto the browser in different ways. By
updating HydrateOnce to remove placeholders I appreciate I'm conflating the
pattern with Portals, but this solution felt the least risky. Using a framework
to manage hydration/portals might help solve this kind of problem more elegantly
in the future.