Bump @guardian/braze-components to 6.0.0
This release requires that the BrazeEndOfArticleComponent is supplied
with click handlers from the platform to track clicks in Ophan and
Braze.

Also includes a little bit of refactoring: if dataFromBraze (the extras
data) isn’t available, then we’re never going to be able to render
anything as we won’t know what to render. So if it’s not defined, return
false from the can render function (which allows us to guarantee it’s
there in the render function). We were kind of side stepping this before
because CanShowResult was parameterized with any instead of Meta.