feat: Add server-side logging of events
We have code that has default fallback when an unexpected input is
passed through. Because of this, we have no visibility on errors.
The only way to gain awareness is by logging a warning somewhere.

Building on top of the existing logs.guardianapis.com/log endpoint,
we create a general-purpose method to log non-fatal errors.

Used in RewrappedComponent to gain visibility of cases when we fail
to rewrap the component.

Add a mock to intercept log calls during tests.

We need to set up some further tooling to make sure that the logs are
being looked at.