feat(cmp): add a dummy for the server bundle
The CMP lib code is not server-safe, as it relies on `window`.

We need a way to make sure this code does not run on the server,
so this aliases to a dummy CMP that does nothing. This approach can be
pushed upstream so the CMP exports this dummy directly.