feat: remove need for Provider to useAB
A hook which returns the AB Test Api when available,
or undefined otherwise.

Leverages an immutable SWR to satisfy all requests to the
AB Core. As soon as the tests are available, all instances of
the useAB hook will render.

Previously, we had an AB Provider wrapped around App.tsx.
With Islands, each island need to provide its own Provider,
leading to much prop repetition.

Co-authored-by: Olly <OllysCoding@users.noreply.github.com>