Make useBraze return a BrazeMessagesInterface, not a Promise
We found that for components which consume useBraze, every render was
getting a new promise from useBraze, only one of which was ever
resolving. This was causing Braze messages to not be rendered correctly.
We've simplified useBraze to return either a BrazeMessageInterface or
nothing, instead of a Promise, which has fixed the issue.

Co-authored-by: Oliver Lloyd <oliverlloyd@users.noreply.github.com>