Wire Braze content cards with profile notifications
This allows our marketing colleagues to configure campaigns/canvases in
Braze to deliver content cards to users which will show up as
notifications in the My Account dropdown.

DropdownLinks now have an ID, we match this against the "target" of the
content card to know where to render the notification.

Content cards are controlled with a new switch "brazeContentCards".
Since this integration has only been done on DCR, we don't want to start
using it just yet as users could get a weird experience where they'd
have notifications on DCR rendered pages but not on frontend rendered
pages.

Co-authored-by: Tom Wey <tom.wey@theguardian.com>