Load interactives on page idle
Since 10th Match, interactive elements have been configured to load only when they become visible in the viewport:

https://github.com/guardian/dotcom-rendering/pull/4077/files#diff-bfe01c729b02a7387fbe09422851114b59f1445c1785129940980a3d0aaebb16R367

This makes sense from an efficiency standpoint but presents issues for spacefinder which has to wait for all interactives to load before running. This is because interactives have a variable height which cannot be anticipated in advance, and spacefinder needs to know this to make sensible decisions about where to place ads.