Refactor second row
Removes the imported fixed/small/slow-iii component for the second row.
Replaces it with the component markup from that component instead.
The only change made to the markup should be that the headline size is
set to 'small' in this file, but not in the slow-iii component. This
reflects the existing designs on Frontend.