Remove `customStyles` prop
Removing `customStyles` from `LI`, and removing the use of it in
`ExtraCardsContainer`, in preparation for a different solution to the
problem of offsetting the verticalDivider length.

We decided that using a css variable which was set in the container
component to adjust the divider length in the child `LI` components
added too much opacity to the rules determining the output.