Show user message when error received
If Frontend returns an error, display a message to the user.

nb. I've tried to keep the state of this component simple, but if higher
priority is put on the design or UX of this component, introducing more
state (i.e. tracking errors in state) would make it easier to style the
message, and also to automatically reset the button state to 'More' if
an error prevents cards from being loaded.

Co-authored-by: James Gorrie <jamesgorrie@users.noreply.github.com>