Consolidate hasMore and collectionId props
Both hasMore and collectionId were only introduced into
containers to support the functionality of ShowMore.importable,
and because the two properties are linked insofar as the collectionId is
only used when hasMore is true. To simplify the interface, following a
suggestion from Oliver Lloyd, this commit consolidates the two props
into one: `showMoreId`. This prop will only be given a value when there
are more cards to fetch for the container, so it can serve both to
provide the id for the API request and to indicate the status of `hasMore`.

The functionality of the `.stories` files should remain unchanged after
this refactor.