Use an article that will expire on 31/12/2022 for paidContentPage in paid.content cypress test
Paid content can be removed in articles used for this test and it starts failing. We have asked for Commercial to provide us with a permanent test page. Until then we're using the URL with the longest end date.

Also removing a comment that is no longer relevant. After this PR https://github.com/guardian/dotcom-rendering/pull/4187 we can do cy.visit(url, { headers : { 'x-gu-geolocation': 'ip.dummytext,country:UK' } }) and the article edition will default to the value in the header. We don't need to do this with this new article we're using because this paid content has the same commercialProperties for all regions.