Cards: Only set `discussionId` if `isCommentable`
Set the value of `discussionId` in the `enhanceCards` function for a
card only if `isCommentable` for that card is true.

This means that the `CommentCount` component will no longer be rendered
on cards which don't have a discussion to display.

There are some potential edge-cases where discussion might have been opened and
then closed, so there might be non-zero comment counts which we don't
display. However, because the discussion section on articles is also
controlled by `isCommentable`, the edge cases shouldn't be visible, and
if we did want to show comments and comment counts for these cases, this
seems like a separate issue.