fix: vertical divider touching last card
When a last row has cards that stretch
to fit the number of columns, the previous
row should not be padded.

The fix consists in only considering the cards
that are forming full rows to pass as an argument
to `shouldPadWrappableRows`.

Using `length - (length % columns)` always ignores
cards from incomplete rows.