Remove ExtraCardsContainer
The `ExtraCardsContainer` component was created as part of the work to
introduce `show more` functionality to DCR Fronts in [#5273](https://github.com/guardian/dotcom-rendering/pull/5273).

The work on `show more` is currently blocked, so this component is not
being used. In addition:

1. The requirements for this component seem likely to change (the
   existing implementation was shaped by the need for client-side
   deduplication of cards, which is likely to be deprioritised for now).
2. Since this component was merged, a number of new patterns have been
   added to the DCR codebase relating to containers and slices, and it
   would be good for the eventual implementation of `show more` to be
   consistent with these new patterns.

Given this, and given that unused code adds maintenance overhead, it
seems to make sense to remove this code for the time being, and to
recreate the container using up-to-date patterns when work on `show
more` starts again, and when the requirements are more clearly
understood.