Move isDev and enableSentryReporting checks
A 'beforeSend' hook was being used inside the Sentry options to check whether
`isDev` or `enableSentryReporting` were true.

These values are available prior to Sentry being loaded at all, so this
moves the check earlier in the process. (There doesn't seem much point
in loading the code if errors are never going to be sent.)