Refactor `oneBig` SecondSlice (remove `shouldPadWrappableRows`)
Refactor
`Card25_ColumnOfCards25_ColumnOfThreeCards25_ColumnOfThreeCards25`
so that it doesn't use `shouldPadWrappableRows`.
We know that the right hand column should consist of three rows of
three, three and two columns respectively. The desired layout can be
created with three static 'mini-slices', instead of a single wrapping
list. (The latter requires complex logic about when to pad the bottom of
wrapped cards.)