Update id for show more container selector
Not all containers have an id of the form `container-${container.id}`,
which is added for the show/hide container functionality.

But all of them should have a div with an id of `container.id` added by
`ElementContainer`. So we're using that id to select containers for show
more instead.

Co-authored-by: Ashleigh Carr <ashcorr20@gmail.com>
Co-authored-by: Georges Lebreton <georges.lebreton@guardian.co.uk>