Remove 'show more' subheading
Semantically the cards in the 'show more' container are part of the same
section as the rest of the cards for that container, so it doesn't work
to have a sub-heading attached to these cards.

The aria description for the show more button should help to orient
screen-reader users to the shift of context, and the heading for the
overall container still applies.