Log header notifications to Braze (#6626)
This logs header notification impressions to Braze. The impression event is
sent when the My Account dropdown is expanded and the relevant notification is
in the viewport (using the useIsInView hook).

We do this so that the Braze canvas which the notification came from can send
users down different paths depending on whether they saw the notification.