Make `heading` a required property on TrailTabType
Currently the only time `heading` should be `undefined` is if it's set
as `undefined` intentionally in `FrontMostViewed.tsx`. The purpose of
this is to signal that the tab heading shouldn't be shown, but this
rule can be enforced using a more transparent test at the point of
rendering in MostViewedFooterGrid, by checking how many tabs there
are.

Co-authored-by: Ioanna Kokkini <ioannakok@users.noreply.github.com>