Remove test: heading render for single tab
This test case was faulty. It tested for a case which should be present on
Fronts, namely that the MostViewed component was passed only a single
'tab' to render.
The case passed even though Fronts were not passing the test in
production. This is because the header-rendering behaviour ins
MostViewedFooter is controlled via the `tab.heading` property, which is
sometimes set to `undefined` by the `FrontMostViewed` wrapper used in
Fronts. Because this test is written using a different wrapper for the
MostViewedFooter component (i.e. MostViewedFooterData.importable.tsx),
which always passes a defined header to its children, the test case was
passing even though the intended behaviour isn't being manifested.
I'm removing this test, instead of refactoring it, because it's not
obvious that the new behaviour that the test was checking for was ever
actually present in `main`. It's unclear at present whether the new behaviour
should be re-implemented or not, so we're sticking with the status quo
behaviour and defer writing a new test case until we've decided what the
desired behaviour is.

Co-authored-by: Ioanna Kokkini <ioannakok@users.noreply.github.com>