Send VIEW and INSERT events to Ophan for header notifications (#6738)
This PR adds Ophan tracking for header notifications (INSERT and VIEW events).
In both cases, a tracking event will be sent for the menu item only if it has
any notifications.

The component event type is RETENTION_HEADER. The component ID is the unique
identifier for the menu item (e.g. account_overview, settings etc.) The
specified notifications will be reflected in the labels property sent to Ophan
(the values sent here will come from the content card data sent from Braze, so
are configurable).