Distinguish `collectionId` from `containerId`
- Rename the `containerId` prop to `collectionId`.
- Add a new prop, `containerElementId`.
- Update usages of the `ShowMore.importable` component accordingly.
- Add documentation of these properties.
- Pass-by refactor of the story -- having more props now makes it harder
  to confirm visually that the two stories are using the same props.

Previously, `containerId` was being used for two purposes:

1. to build the URL for the `show-more/` API endpoint
2. to select the main container for the collection on the client side

As of [#7108](https://github.com/guardian/dotcom-rendering/pull/7108)
the container element's id is being set to the `ophanName`, meaning we
couldn't use the same prop value for both of the above purposes, and the
API call was failing because it was using the wrong id value. This PR
distinguishes two separate props for the two roles, and documents the
differences between them.