Change how story names are set (#7591)
Some stories have an object in their `.story` field with multiple properties, one of which is the `name` property. This moves the `name` property to a `.storyName` field directly on the story object. The previous way doesn't match the documentation and doesn't seem to work in storybook 7.

Having other properties in the `.story` object does seem to work in storybook 7, so these have not been changed.