Dependabot: ignore @types/node major & minor
For DefinitelyTyped types, the major and minor versions of the type
package should match the major and minor versions of the package being
typed. Dependabot knows this for packages where the main package is
installed via npm, but it doesn't know this for node itself.

This should stop Dependabot from raising unneeded PRs for `@types/node`.