Introducing a sponsor badge for the Guardian Labs container.
Collections already have the notion of a badge, currently in order to display a badge for the "This is Europe" container.

A collection will also display a badge if it is a labs collection and all the cards in the collection have the same sponsor. In this case it will be that sponsor's logo.

wip: isLabs property