Add a new publish workflow.
This new workflow will extract the server files from the prebuilt container and upload it to riff-raff for deployment.

Use a tarball instead of a zip file for publishing DCR files to the riffraff bucket.

zip / unzip is generally not included in a lot of images, especially alpine ones. Meanwhile tar is widely available on most Linux distros.

Add buildNumberOffset to actions-riff-raff and comment explaining why we're not uncommenting

Remove old generation of build.json. This is handled by guardian/actions-riff-raff now.

Co-authored-by: George B <705427+georgeblahblah@users.noreply.github.com>
Co-authored-by: Ashleigh Carr <ashcorr20@gmail.com>
Co-authored-by: Parisa Tork <parisa_tork@yahoo.co.uk>
Co-authored-by: Daniel Clifton <110032454+DanielCliftonGuardian@users.noreply.github.com>
Co-authored-by: Sophie <26366706+sophie-macmillan@users.noreply.github.com>
Co-authored-by: Anna Beddow <anna.beddow@gmail.com>
Co-authored-by: Charlotte Emms <43961396+cemms1@users.noreply.github.com>