refactor(Doughnut): arcs for segments
Using dash arrays on circles leads to visual rendering
bugs in Chromium, so this refactor introduces actual
SVG args to represent segments.

A circle is still used for a single segment as arcs cannot
have the same point for start and end.

the SVG viewBox is now centred to 0,0 making some calculations
simpler, as points now lay perfectly on the trigonometric circle.