Add z-index: 1 to the VeggieBurgerMenue
To make sure part of it does not get hidden. This is to address an edge case of when the first container is a thrasher and the front path starts with `news/series` or any other combination that does not open any submenu. (On the other hand, a path that opens the News submenu would be `uk-news/series`.). Screenshots are added in the PR.