feat(performance): lower TTFB threshold
We want to improve sensitivity with minimal cost to specificity.

That means correctly identify more pages that will actually be slow,
while keeping the number of pages that are fast and incorrectly
identified as slow minimal.

For the last two weeks of page views, here’s how that breaks down:
- TTFB 1200+ : Sensitivity 7.43%, Specificity  99.23%
- TTFB 2400+ : Sensitivity 3.33%, Specificity 100.00%