feat: support Australia state switcher
If a container is targetted to a specific Australian territory,
a switcher is inserted below in place of “Show More +” to allow
users to change the localised news they are presented.

Compared to Frontend, we removed “Hide local news”

Hiding the container is sufficient for this use-case,
and when people click this button they cannot get
the local news container ever again–bar modifying
their cookies manually.