Remove `atoms-rendering` From AR (#7958)
* Removed `atoms-rendering` from AR

It's not used for any of the atoms AR has available for production users (media atoms, special report alt atoms), and we're not going to add support for the atoms it is used for. Removing the dependency simplifies the maintenance of AR.

This change removes all rendering code that relies on `atoms-rendering`, and removes the dependency. It doesn't remove the parsing code that extracts atoms from the CAPI response, as this doesn't rely on `atoms-rendering`, would be more complex to remove, and might still be useful in DCR at some point.

* Updated renderer tests

Make sure unsupported atoms render `null`