feat(Card): Support mainVideo
The settings comes from the facia-tool via Frontend,
when “Show Video” is enabled for a trail in a container,
it should be displayed as a playable video inline instead
or redirecting to the article itself.

Currently, only Youtube Atoms are supported.

There is currently no guarantee that the first atom will be the expected
video, but it happens to be the case the vast majority of the time

A minimum Card width of one third or three columns is required for
the YouTube atom to be embedded, similar to the existing Frontend logic:
https://github.com/guardian/frontend/blob/1c4555a9f/common/app/layout/cards/CardType.scala#L20-L23