Modify kicker logic
Now that we're rendering both pressed and tag fronts we need to modify the kicker logic. In pressed fronts, we should show thw `kickerText` if it has been provided by the editors in the fronts tool. In tag fronts, we should show the `seriesOrBlogKicker` if present, only if it's not a series tag front.

Co-authored-by: Olly Namey <olly.namey@guardian.co.uk>
Co-authored-by: Charlotte Emms <charlotte.emms@guardian.co.uk>