Used full paths to source directories over `cwd`
The way these calls to `cpy` used to work, setting `cwd` and then using a glob to specify paths in that directory, no longer works with the updated version of `cpy`. This rewrites these calls so that they specify the full path in the `source` argument, rather than using `cwd`.

It also splits the code to copy the `stats` directory into a separate call as, again, the previous way this was done no longer works.