refactor(enhancers): use Array#map where possible
Array.prototype.map is more declarative and nearly
as performant as an imperative for loop.

We can be certain that the result of this computation
is an array of the exact same length. The compiler
can give us hints if we forget to return a value
in any iteration’s branch.

No further functional refactors have been performed
on the internals of these loops. Tests are left as-is.