fix: change deferUntil to visible for AustralianTerritorySwitcher
There seems to be a bug with the `deferUntil="interaction"` as seem on the live site.
This has been reproduced, but is intermittant and I can't figure out what the conditions are.

While this seems like the right `deferUntil` - I suggest we change it to visible while we try to work out the bug.

It hasn't been reported on any other Islands with the same deferUntil.