Skip Cypress tests in TeamCity if `SKIP_CYPRESS` is set (#8451)
* Run Cypress tests if `SKIP_CYPRESS` is not set

We've seen a lot of failures of our Cypress tests in our TeamCity builds recently. Cypress tests are required to pass for the build to succeed.

When we have changes that need to be released quickly and the cypress tests are failing, this can become a blocker.

Typically, we either retry or disable cypress tests in TeamCity temporarily (as in #8447). Although this works, it can be a slow process to raise the PR and wait for the builds.

Until we have a robust solution to fix the Cypress tests, we can disable
the tests by setting this environment variable to any value.

Resolves #8450

* Explain how to bypass Cypress tests if they fail