Add an apps footer to `StandardLayout`
The apps footer shows some copyright information as well as buttons to
privacy settings and the privacy policy. If the user is affected by CCPA
(California Consumer Privacy Act), we include a button to access the privacy
policy with some different wording.

When the buttons are clicked, we call Bridget to show the relevant
information in the native UI.