DCAR: Follow button (#8570)
If we are rendering for apps and the article has a single ("sole") contributor, we should show a button so the user can toggle the follow status.

If:
- the user is not following the contributor, we should show a "Follow" button 
- the user is already following the contributor, we should show an "Unfollow" button

Additionally adds a polyfill for `Buffer` in storybook. I don't _think_ we need to polyfill this for browsers, but maybe we do. Apps Rendering adds a polyfill directly to the Webpack config as well as Storbyook.

---------

Co-authored-by: Jamie B <53781962+JamieB-gu@users.noreply.github.com>
Co-authored-by: Parisa Tork <47482049+ParisaTork@users.noreply.github.com>
Co-authored-by: James Gorrie <jamesgorrie@users.noreply.github.com>
Co-authored-by: Ravi <7014230+arelra@users.noreply.github.com>