Fix Editions avatar background colour in dark mode (#8721)
The Editions app doesn't support dark mode, so everything should look the same whether the user is in light or dark mode. There's a bug where the contributor avatar that appears on Comment pieces was switching to a different background when in dark mode. This happened because the underlying `Img` component Editions uses is shared, and comes with a dark mode background. The fix is to add some dark mode styles specifically for Editions that override this dark mode background, setting it to `none` instead.