Add Standard article fixture for Cypress
The fixture we were using has previously only been used for Storybook
tests, meaning it isn't used to render and test a full interactive page
render. Some parts of this Storybook fixture are out of date, including
the Commercial Bundle URL. When using this fixture in Cypress, we see
errors and test failures because of this.

This change adds an up to date Standard article fixture for Cypress.