Add `DCRArticle` type to model an enhanced article
We use `FEArticleType` to model data generated by Frontend and passed to
DCR. However, we also add fields to this type which _don't_ exist in the
data sent by Frontend, like `tableOfContents` and `imagesForLightbox`.
This is confusing!

This change adds a `DCRArticle` type which is intended to model an
article type for DCR to use after we enhance the incoming data from
Frontend. We can add fields to this type that do not exist in Frontend.