Enhance Fixtures (#8913)
As explained in https://github.com/guardian/dotcom-rendering/issues/8899, we use fixtures as input data for stories and unit tests. However, the fixtures are of type `FEArticleType` which models the data we receive from Frontend. When handling a request, DCR "enhances" this data, converting it to a `DCRArticle` (https://github.com/guardian/dotcom-rendering/pull/8881). This change modifies the fixture generation to also enhance `FEArticleType` data, so that the fixtures reflect the shape of data that the application actually sees.