`useMatchMedia` for ads
On narrow breakpoints we want ads to appear in line in the body of the article, and on wider ones we want them to appear in the right column. When the user transitions between breakpoints (e.g. rotates their device or resizes in some other way) we want the native ads to be re-drawn in their new positions: right column to inline or inline to right column.

To observe breakpoint changes in JavaScript, `window.matchMedia` and an event listener can be used. This change creates a custom hook, `useMatchMedia`, to surface the match status of a given media query to components. Any change will trigger a re-render and provide a new value.

In the case of `AdPortals`, the re-render will move ads between their inline and right column slots. This will in turn change the height of the page and so trigger a Bridget call to update the native ad positions.