Use apps algorithm for liveblog and deadblog ads
Modifies the insertion of ad placeholders in apps blogs, to ensure ad parity with existing AR articles.

The ad placeholder is now used in a few places, so this change moves it into a separate component for reuse. It also exports the class name for use in `AdPortals`, to ensure the client-side code is always looking up the same class that's provided on the server.