Add `showMainVideo` flag in the `DCRFrontCard`
The flag is set in the fronts tool and is sent to DCR in `faciaCard.properties.showMainVideo` JSON property. We need to know if this is set in the `YoutubeBlockComponent` to be able to show or hide the play button and the duration of the video. Currently we show those regardless off what the value of the flag is.