Respect `showMainVideo` flag in media of type `picture`
The `showPlayIcon` boolean is true when the mainMedia is of type `Video` but `isPlayableMediaCard` is false. The size of the card determines whether a media card can be playable or not because YouTube recommends a minimum width of 480px (https://developers.google.com/youtube/terms/required-minimum-functionality#embedded-youtube-player-size). If a card is less than 480px we render the image using the `CardPicture` component and not the `YoutubeAtom`. This commit makes sure we respect the flag in both cases.