Stop sending Okta errors to Sentry (#9025)
Since rolling out Okta to 100% of readers, we have seen a spike in dropped errors in Sentry. Dropped errors are bad because it means we might lose sight of errors we care about. This change stops sending errors originating from the Okta flow to Sentry, instead just logging them to console.