Use `overrideImage` first and fallback to `posterImage` if needed
When `showMainVideo` flag is off we want to render the picture using the `CardPicture` component. After https://github.com/guardian/dotcom-rendering/pull/9012 the trail image (`overrideImage`) takes priority over the video thumbnails (`posterImage`). If it doesn't exists, I am assuming that the image with the largest width from `posterImage` is the `mainImage` we want to use.