Decrease `minCapacity` of instances to 25
`minCapacity` was bumped from 15 to 30 in https://github.com/guardian/dotcom-rendering/pull/8724. After the fronts migration was completed, it seemed that 15 instances weren't holding up with the new traffic we were sending. That was a hotfix and there is upcoming work to split our stacks (https://github.com/guardian/dotcom-rendering/issues/8351) but until we reach this point we would like to see whether we could handle the traffic with 25 instances. This change is part of the research we're doing to optimise how we scale: https://github.com/guardian/dotcom-rendering/issues/9322.

Co-authored-by: George B <705427+georgeblahblah@users.noreply.github.com>
Co-authored-by: Ravi <7014230+arelra@users.noreply.github.com>