Simplifies `Parser` type in the `parser` module
A `Parser` is a function from `unknown => Result`, there's no need to wrap it in an object with a `run` method. Removing this extra layer should make the code easier to read. It also means prettier no longer tries to add so many new lines, so the formatting is more compact.