Dev Server Article Examples Table (#9614)
It's often useful to have examples of articles to work with when making changes. This creates a way to add new examples and view them across our different products and environments. It also displays the format of each example. Only two examples are included for now, and more can be added later.

The rows and the links for different products and environments are built using web components, rather than a framework like React, as the dev server uses JavaScript without any dependencies.

Note that this table doesn't currently work in Safari, as customising built-in elements is not supported there. This will be fixed in a future change.