Refactored the article examples table
The first benefit of this is that it now works in Safari. Previously it made use of custom elements that extend built-in elements, which Safari doesn't support. Now the whole table is wrapped in an autonomous custom element instead. This was also an opportunity to create a Shadow DOM for the table, to isolate the styles from the rest of the page.

Another change is what the article examples display when JS is not available. Previously this would have been nothing, but now it's a list of links to the production dotcom version of the articles. This means the list is still useful without JS, and the custom elements behaviour just adds functionality.