Add Support For Descriptions Of Article Examples (#9806)
It's useful to be able to add more information about an article in the example list; for example, what elements it contains. The structure of this information is likely to vary based on the article, and HTML provides more formatting flexibility than strings.

This adds the ability to include a "slot" in each list item with an HTML description. This maps onto a `slot` element in the row template, which is populated with this HTML.

Although this mechanism uses the same semantics as the "slotting" feature of web components, like `slot` elements and attributes, it does _not_ use that feature directly. It's not possible to have slots that are not the direct children of a custom element, and the `tr`s cannot be made custom elements as this is not supported in Safari. As a result this change manually performs similar functionality by moving the descriptions from the custom element descendants into the slots in the shadow DOM.