Fix rendering of notification badge in expanded dropdown (#10131)
The styles were changed in #7671 to increase the clickable area of each
menu item. But the removal of a "display: flex" had the effect of
causing the badge for a menu item (when expanded) to appear below the
item instead of to the right as intended. This fixes the issue while keeping
the expanded clickable area.

Co-authored-by: Max Duval <max.duval@theguardian.com>