Examples for managing Discussion with `useState` and `useReducer`
`DiscussionServerSide.tsx` is the component to be rendered on the server. It chooses one of two islands to render depending on the platform.

`DiscussionApps.importable.tsx` and DiscussionWeb.importable.tsx` are the islands for each platform. They contain examples of the different platform-specific effects that are passed to the main discussion component. The effects have identical types, so the main discussion component doesn't know or care which set of effects have been passed to it.

`DiscussionWithUseReducer.tsx` and `DiscussionWithUseState.tsx` represent the main discussion component that handles state and effects for discussion. Two are included for example purposes, but in practice there would only be one of these, and it would probably just be called `Discussion.tsx`.

`Comments.tsx` contains presentational components that display the comments, forms and error messages.