URI encode branding `src` to make sure it's valid (#10365)
Branding asset URLs sometimes include spaces, which are not valid URI characters. This doesn't necessarily cause a problem when used in a `src` attribute, as browsers can encode them automatically. However, in the `srcset` attribute, whitespace is used to separate URLs and condition descriptors, so it doesn't automatically get encoded.

In most cases in light mode, an `img` with a `src` attribute is used for branding logos. So if the asset URL contains spaces it is still displayed correctly. However, in dark mode, and in some cases on liveblogs, a `source` with a `srcset` attribute is used, so an asset URL with spaces will not display correctly.

This change uses `encodeURI` for all logo assets, in light and dark mode, to make sure they display correctly.

- https://developer.mozilla.org/en-US/docs/Web/API/HTMLImageElement/src
- https://developer.mozilla.org/en-US/docs/Web/API/HTMLImageElement/srcset
- https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/encodeURI