fix(Lightbox): build images once
an initial array of `ImageForLightbox[]`
is generated from Main media elements and blocks.

The blocks and main media are then enhanced with `position`
if the images is also found in the list of images for lightbox.

the presence of `position` is now the only necessary indicator
of what makes an image “lightboxable”