refactor(Comments): clearer comment batching
Improves reliability of scrolling to a specific comment

The previous behaviour did not reset when the comments
array is modified, so any rendering benefits are only present
on the very first render. It also did only two batches instead
of incrementally reaching the final total number of comments.

The number of comments (10) is now extracted to a constant
that is well documented to prevent accidental shift in the future.